requirements.txt
numpy-1.23.4
pandas-1.5.0
#matplotlib-3.6.0 
#seaborn-0.12.1
tensorflow-2.10.0